+++ tests/src/smtp/outbound/tls.rs
@@ -144,7 +144,9 @@ async fn starttls_optional() {
     let next_due = now();
     let queue_id = retry.queue_id;
     retry.message.recipients[0].retry.due = next_due;
-    retry.save_changes(&local.server, prev_due.into()).await;
+    retry
+        .save_changes(&local.server, prev_due.into(), None)
+        .await;
     local
         .delivery_attempt_for_queue(queue_id, "default")
         .await